+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - http optimistic concurrency control
 - longer anonymous cookie max_age (one month instead of session only)
 - add possibility to override default zoom with LEAFLET_ZOOM setting
+- fix bug where anonymous map wasn't editable by logged in users even if
+  edit status was ANONYMOUS
 
 ## 0.5.x
 - internal storage structure totally reviewed: datalayers are stored as geojson files,